# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "description": "Description of your API",
        "contact": {
            "name": "Contact Name",
            "email": "contact@example.com"
        },
        "license": {
            "name": "License Name",
            "url": "https://www.example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/carousels": {
            "get": {
                "tags": [
                    "Carousels"
                ],
                "summary": "Get list of carousels",
                "operationId": "cff0a0e73b98416d688570a17c5e2aaa",
                "responses": {
                    "200": {
                        "description": "List of carousels"
                    }
                }
            },
            "post": {
                "tags": [
                    "Carousels"
                ],
                "summary": "Store a new carousel",
                "operationId": "7e2e73ff325a73fced463b87279e91d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "desc",
                                    "img"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Carousel Title"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Carousel Description"
                                    },
                                    "img": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carousel created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/carousels/{id}": {
            "get": {
                "tags": [
                    "Carousels"
                ],
                "summary": "Get details of a specific carousel",
                "operationId": "9b80b2e90ecade0155d25ff2df06334d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the carousel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carousel details"
                    },
                    "404": {
                        "description": "Carousel not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Carousels"
                ],
                "summary": "Update details of a specific carousel",
                "operationId": "697ff719be25699b8fffa5d40cb03e29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the carousel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "desc",
                                    "img"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Carousel Title"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Updated Carousel Description"
                                    },
                                    "img": {
                                        "type": "string",
                                        "example": "Image Title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carousel updated successfully"
                    },
                    "404": {
                        "description": "Carousel not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carousels"
                ],
                "summary": "Delete a carousel",
                "operationId": "21ba2ce7e091a1d347dbd07ec67498e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the carousel to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Carousel deleted successfully"
                    },
                    "404": {
                        "description": "Carousel not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get list of categories",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "List of categories"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create a new category",
                "operationId": "createCategory",
                "requestBody": {
                    "description": "Category data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get a specific category",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update a specific category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Category data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete a specific category",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    }
}